Supporting extension x-okta-links
First pass at the logic of _moving_ a method from the ‘API’ client to the model.
Might need to rename packages as the models are now more resources then models.
The clients should get trimmed down automatically when the swagger spec is updated and more operations are moved the respective models