IDEA: move config parsing from spring module into the api module for reuse
Was thinking about how we could support Spring, Guice, and CDI with some common code.

Basically move the logic of parsing a configuration object into the core code.
The way to pull in configuration differs between DI
- Spring - you can it is a trivial annotation
- with Guice and CDI, there are many different ways to do this and no real standard
- CDI has NO way of doing this with out an extension, and it isn't the job of this SDK to force dependencies on someone

Each implementation allows for the optional injection of a CacheManager and ClientConfiguration

Few warts are in here, but should be good enough for discussion
- the ClientConfiguration classname conflicts with the implementation in the DefaultClientBuilder
- ClientConfiguration's nested structure comes from the way things work with Spring @ConfigurationProperties works