Create RetryRequestExecutor
Port retry logic from HttpClientRequestExecutor to new RetryRequestExecutor

Use new RetryRequestExecutor as parent for HttpClientRequestExecutor

The RetryRequestExecutor is meant to delegate to a RequestExecutor, but in this case it is used as the parent to keep backward compatibility.
This also means that all of the deprecated bits from HttpClientRequestExecutor are moved into RetryRequestExecutor.