Scope Handlebars dependencies to openapi generator plugin
The built artifacts do not depend on Handlebars (at least not according
to `jdeps`), nor do any of this project's other dependencies.

The version of Handlebars required by this constraint is only compatible
with JDK 17 and upwards; this project would otherwise be compatible with
earlier JDKs. Since the dependency is actually unused by the Okta SDK, the
incompatibility with JDK 11 is a bit subtle: it will only manifest in
applications that are themselves using Handlebars at runtime.